pyproject.toml
[project]
name = "dage-gpu"
version = "0.1.0"
description = "Multi-GPU LLM LoRA Training with DeepSpeed + Zeus"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 核心训练框架
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    # LoRA 参数高效微调
    "peft>=0.7.0",
    # 其他
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
    "tensorboard>=2.15.0",
]

[project.optional-dependencies]
# GPU 训练依赖 (仅 Linux)
gpu = [
    "deepspeed>=0.12.0",
    "bitsandbytes>=0.41.0",
    "zeus-ml",
]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
